+++ Rollbar/RollbarResult.cs
@@ -1,0 +1,9 @@
+namespace RollbarDotNet
+{
+    public class RollbarResult
+    {
+        public int? Id { get; set; }
+
+        public string Uuid { get; set; }
+    }
+}
